cube가 제대로 렌더링되지 않는 현상 수정 완료, cube를 로드할 때 position만 가져오고 나머지 데이터는 가져오지 않도록 수정, 모든 object가 color를 가지고 있도록 설정

+++ SpaceshipGame/ModelLoaderClass.hpp
@@ -42,7 +42,7 @@ void Graphic::Loader::ModelLoaderClass<VertexType>::LoadVertex(const aiScene* sc
 	VertexType vertex = {};					// vertex 데이터
 
 	// 메쉬 개수를 가져와서 가져온 메쉬 개수만큼 데이터(vertex, index, texture) 파싱 //
-	for (ULONG i = 0; i < GetMeshCount(); ++i)
+	for (UINT i = 0; i < GetMeshCount(); ++i)
 	{
 		mesh = scene->mMeshes[i];
